% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{invlogit}
\alias{invlogit}
\title{Function to calculate inverse log odds}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A numeric vector of negative or positive values}
}
\value{
A vector of probabilities
}
\description{
Function to calculate the inverse log odds also known as inverse logits.
}
\examples{
x <- rnorm(10)
invlogit(x)
}
