% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{logit}
\alias{logit}
\title{Function to calculate log odds}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric vector of probabilities}
}
\value{
A numeric vector of the log odds
}
\description{
Function to calculate the log odds, also known as the logits.
}
\examples{
p <- c(0.25, 0.5, 0.75)
logit(p)
}
