% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testR-package.R
\docType{package}
\name{testR-package}
\alias{testR}
\alias{testR-package}
\title{testR: A Test Package By Lucy}
\description{
Functions made by Lucy to add to a package. Functions include 
    calculating log odds and inverse log odds.
}
\author{
\strong{Maintainer}: Lucy Justice \email{lucy.justice@ntu.ac.uk}

}
\keyword{internal}
